Act on blank option types table and adjust Grid query test
In subsequent review of the observer code, it was noticed that if the products_options_types table were emptied, that the Grid product option type would be added as 1 which is a pre-existing ZC defined value.  Instead, if this is added into a cart where the products_options_types table has been cleared, then the initial value for it will be set to 6 which is the first number after the ZC standard/default listing for products_options_types product types.
Also the check performed for the presence of Grid being already in the database did not use a standard query check of the associated data.  This was modified to be more consistent with routine processes.