Permit SBA Stock to permit display of attribute quantities.
By default, ZC product types permit the display or hiding of the total quantity of product available. The Grid option type offers the ability to display the quantity remaining of an attribute "combination"; however, it was dependent on the product types permission to display quantity available.  SBA offers a separate quantity display control related directly to the attributes themselves.  Therefore, this additional control/permission is added and could be modified to support the needs of other such plugins.