Show single option values in PAG when PAG displayed.
Fixes #3.

This makes it possible for PAG to display the attribute name in the PAG list if there is one and only one option value for the applicable option name and if the display of single option values is disabled in the attributes.php file.  More is needed to correct the issue if there is one option value and the attributes.php file has not been modified.