Remove excess radio button when single option value
It was identified that when an option name had a single option value that in some way both the product attribute grid and the single radio button were incorrectly mixed together for presentation. The radio button should not be displayed at all.  Because the radio button is determined necessary from within the includes/modules/YOUR_TEMPLATE/attribute.php file prior to/independent of the option name type, this extra button needs to be removed from the generated array for attributes.  This commit performs that.