Force display of identified attribute id
Although ZC is basically designed to not have an option name, attribute_id, etc... with a value of 0, the code permitted some of those values to be assigned as blank which if using regex or some other format comparison tool may result in unusual results.  This commit forces a value of 0 to be displayed instead of nothing.